% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_body.R
\name{check_bullets}
\alias{check_bullets}
\title{Insert bullets before text}
\usage{
check_bullets(text)
}
\arguments{
\item{text}{output from extract_text or stylyize_text. Typically the latter.}
}
\description{
Insert bullets before text
}
\keyword{internal}
