% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_lang}
\alias{check_lang}
\title{Check the Language of PPTX}
\usage{
check_lang(xml_folder)
}
\arguments{
\item{xml_folder}{The folder containing all of the xml code from the pptx,
created from \code{\link{extract_xml}}.}
}
\description{
Check the Language of PPTX
}
\keyword{internal}
