% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{convert_pptx}
\alias{convert_pptx}
\title{Extract xml from pptx}
\usage{
convert_pptx(path, author, title = NULL, sub = NULL, date = Sys.Date(),
  theme = "default", highlightStyle = "github", force = FALSE,
  writenotes = TRUE, out_dir = NULL)
}
\arguments{
\item{path}{Path to the Microsoft PowerPoint file}

\item{author}{A string indicating th author or authors of the slide deck.
Multiple authors can be provided with a string vector.}

\item{title}{Optional title of the slide deck. Defaults to the title of the
first slide in the deck.}

\item{sub}{Optional subtitle}

\item{date}{The date the slides were produced. Defaults to current date.}

\item{theme}{The css theme to apply to the xaringan slides. For options, see
\href{https://github.com/yihui/xaringan/tree/master/inst/rmarkdown/templates/xaringan/resources}{here}.
Note that only the name of the theme needs to be applied
(e.g., \code{theme = "metropolis"}) and the proper code will be applied to
load both the theme and the fonts, although this can easily be manually
manipulated after conversion if you want other fonts with a specific theme.}

\item{highlightStyle}{The code highlighting style. Defaults to
\code{"github"} flavored highlighting}

\item{force}{If an 'assets' folder already exists in the current directory,
(e.g., from a previous conversion) should it be overwritten? Defaults to
\code{FALSE}.}

\item{writenotes}{Logical. If notes are in the original PPTX, should they be
written out as their own .txt file in addition to being embedded in the
slides? Defaults to \code{TRUE}.}

\item{out_dir}{Optional output directory for the folder containing the RMD
and corresponding assets. Defaults to the current working directory.}
}
\description{
Extract xml from pptx
}
