% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml.R
\name{create_yaml}
\alias{create_yaml}
\title{Create the \href{https://github.com/yihui/xaringan}{xaringan} YAML Front
Matter}
\usage{
create_yaml(
  xml_folder,
  title_sld,
  author = NULL,
  title = NULL,
  sub = NULL,
  date = Sys.Date(),
  theme = "default",
  highlightStyle = "github"
)
}
\arguments{
\item{xml_folder}{The folder containing all of the xml code from the pptx.}

\item{title_sld}{The xml code for the title slide in the pptx.}

\item{author}{Optional string indicating the author or authors of the slide.
Defaults to the listed creator of the pptx.Multiple authors can be
provided with a string vector.}

\item{title}{Optional title of the slide deck. Defaults to the title of the
first slide in the deck.}

\item{sub}{Optional subtitle}

\item{date}{The date the slides were produced. Defaults to current date.}

\item{theme}{The css theme to apply to the xaringan slides. For options, see
\href{https://github.com/yihui/xaringan/tree/master/inst/rmarkdown/templates/xaringan/resources}{here}.
Note that only the name of the theme needs to be applied
(e.g., \code{theme = "metropolis"}) and the proper code will be applied to
load both the theme and the fonts, although this can easily be manually
manipulated after conversion if you want other fonts with a specific theme.}

\item{highlightStyle}{The code highlighting style. Defaults to
\code{"github"} flavored highlighting}
}
\description{
Create the \href{https://github.com/yihui/xaringan}{xaringan} YAML Front
Matter
}
\keyword{internal}
