% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_author}
\alias{extract_author}
\title{Extract Author from PowerPoint}
\usage{
extract_author(xml_folder)
}
\arguments{
\item{xml_folder}{The folder containing all of the xml code from the pptx.}
}
\description{
Extract Author from PowerPoint
}
\keyword{internal}
