% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_body}
\alias{extract_body}
\title{Extract the body of the slide}
\usage{
extract_body(sld)
}
\arguments{
\item{sld}{xml code for the slide to extract the body from}
}
\description{
Extract the body of the slide
}
\keyword{internal}
