% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_class}
\alias{extract_class}
\title{Extract Slide Element Classes}
\usage{
extract_class(sld)
}
\arguments{
\item{sld}{xml code for the slide to extract the title from}
}
\description{
Extract Slide Element Classes
}
\keyword{internal}
