% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_footnote}
\alias{extract_footnote}
\title{Extract Footnotes from Slides}
\usage{
extract_footnote(sld)
}
\arguments{
\item{sld}{xml code for the slide to extract the title from}
}
\description{
Extract Footnotes from Slides
}
\keyword{internal}
