% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_body.R
\name{extract_indents}
\alias{extract_indents}
\title{Check bullets before text}
\usage{
extract_indents(xml_tibble)
}
\arguments{
\item{xml_tibble}{output from extract_text or stylyize_text. Typically the latter.}
}
\description{
Check bullets before text
}
\keyword{internal}
