% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_attr.R
\name{extract_link}
\alias{extract_link}
\title{Extract Links from the corresponding slide}
\usage{
extract_link(sld, rel)
}
\arguments{
\item{sld}{xml code for the slide to extract the attribute from}

\item{rel}{xml \code{rel} code for the slide}
}
\description{
Extract Links from the corresponding slide
}
\keyword{internal}
