% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_subtitle}
\alias{extract_subtitle}
\title{Extract Subtitle from Title Slide}
\usage{
extract_subtitle(sld)
}
\arguments{
\item{sld}{xml code for the slide to extract the title from}
}
\description{
Extract Subtitle from Title Slide
}
\keyword{internal}
