% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{extract_table}
\alias{extract_table}
\title{Extract tables from slides}
\usage{
extract_table(sld)
}
\arguments{
\item{sld}{xml code for the slide to extract the table from}
}
\value{
a \code{data.frame} with the data from the table. Generally fed to
  \code{\link{tribble_code}}.
}
\description{
Extract tables from slides
}
\keyword{internal}
