% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_title}
\alias{extract_title}
\title{Extract slide title}
\usage{
extract_title(sld)
}
\arguments{
\item{sld}{xml code for the slide to extract the title from}
}
\description{
Extract slide title
}
\keyword{internal}
