% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_notes_xml}
\alias{import_notes_xml}
\title{Import Slide Notes}
\usage{
import_notes_xml(xml_folder)
}
\arguments{
\item{xml_folder}{The folder containing all of the xml code from the pptx.}
}
\description{
Import Slide Notes
}
\keyword{internal}
