% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_rel_xml}
\alias{import_rel_xml}
\title{Import XML \code{rel} Code from PowerPoint}
\usage{
import_rel_xml(xml_folder)
}
\arguments{
\item{xml_folder}{The folder containing all of the xml code from the pptx.
created from \code{\link{extract_xml}}.}
}
\description{
Import XML \code{rel} Code from PowerPoint
}
\keyword{internal}
