% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_slide_xml}
\alias{import_slide_xml}
\title{Import xml code for pptx slides}
\usage{
import_slide_xml(xml_folder)
}
\arguments{
\item{xml_folder}{The folder containing all of the xml code from the pptx,
created from \code{\link{extract_xml}}.}
}
\description{
Import xml code for pptx slides
}
\keyword{internal}
