% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_body.R
\name{level_indents}
\alias{level_indents}
\title{Fixes the bulleting levels in the case of non-standard nesting}
\usage{
level_indents(indented)
}
\arguments{
\item{bulleted}{A dataframe output from \code{insert_bullets}}
}
\description{
Fixes the bulleting levels in the case of non-standard nesting
}
\keyword{internal}
