% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidex-package.R
\docType{package}
\name{slidex-package}
\alias{slidex}
\alias{slidex-package}
\title{This package is designed to extract information from Microsoft
 PowerPoint slides, and then put that information into an R Markdown
 document with a \href{xaringan}{https://github.com/yihui/xaringan} YAML. If
 the xaringan package is also installed, beautiful html slides can then be
 produced by knitting the RMD. At present, the package exports one function,
 `convert_pptx`, which converts a .pptx file to R Markdown. The package is
 not intended to be all encompassing or provide perfect conversion. Rather,
 it should get you about 90% of the way there for about 80% of use
 cases. Given the idiosyncrasies of different slide show presentations, it
 is expected that some manual editing of the resulting RMD will need to be
 completed to get the HTML slides to look "perfect", but most of the hard
 work should be done for you. Importantly, the package maintains and
 provides the proper code for any images that were embedded in the PPTX
 slides, as well as links. Tables are also generally maintained, although
 they may require some manual editing if complex spanner heads and merged
 cells were used in the original table. Nested bulleted lists should also be
 maintained.}
\description{
This package is designed to extract information from Microsoft
 PowerPoint slides, and then put that information into an R Markdown
 document with a \href{xaringan}{https://github.com/yihui/xaringan} YAML. If
 the xaringan package is also installed, beautiful html slides can then be
 produced by knitting the RMD. At present, the package exports one function,
 `convert_pptx`, which converts a .pptx file to R Markdown. The package is
 not intended to be all encompassing or provide perfect conversion. Rather,
 it should get you about 90% of the way there for about 80% of use
 cases. Given the idiosyncrasies of different slide show presentations, it
 is expected that some manual editing of the resulting RMD will need to be
 completed to get the HTML slides to look "perfect", but most of the hard
 work should be done for you. Importantly, the package maintains and
 provides the proper code for any images that were embedded in the PPTX
 slides, as well as links. Tables are also generally maintained, although
 they may require some manual editing if complex spanner heads and merged
 cells were used in the original table. Nested bulleted lists should also be
 maintained.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datalorax/slidex}
  \item Report bugs at \url{https://github.com/datalorax/slidex/issues}
}

}
\author{
\strong{Maintainer}: Daniel Anderson \email{daniela@uoregon.edu}

}
\keyword{internal}
