% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_body.R
\name{stylize_text}
\alias{stylize_text}
\title{bold or italicize text}
\usage{
stylize_text(text)
}
\arguments{
\item{text}{data frame output from extract_text}
}
\description{
bold or italicize text
}
\keyword{internal}
