% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tribble_code}
\alias{tribble_code}
\title{Wrap a DF in \code{tibble::tribble} code}
\usage{
tribble_code(df, tbl_num = "")
}
\arguments{
\item{df}{A \code{data.frame}, typically the output from
\code{\link{extract_table}}.}

\item{tbl_num}{The table number. Not produced in the caption, but used
to name the object and the code chunk. In typical application, corresponds
to the slide number.}
}
\description{
Wrap a DF in \code{tibble::tribble} code
}
\keyword{internal}
