% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_notes}
\alias{write_notes}
\title{Write Slide Notes}
\usage{
write_notes(xml_folder)
}
\arguments{
\item{xml_folder}{The folder containing all of the xml code from the pptx.}
}
\description{
Write Slide Notes
}
\keyword{internal}
