% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml.R
\name{write_theme}
\alias{write_theme}
\title{Write the YAML Code for CSS Theme}
\usage{
write_theme(theme)
}
\arguments{
\item{theme}{The name of the theme, passed as a single string. Note that
just the name needs to be supplied and any accompanying fonts will also
be included.}
}
\description{
Write the YAML Code for CSS Theme
}
