% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_notes}
\alias{extract_notes}
\title{Function to pull notes from a slide}
\usage{
extract_notes(notes, sld_num, inslides = TRUE)
}
\arguments{
\item{notes}{A list of the xml code with all the notes for all slides}

\item{sld_num}{The specific slide number to pull the notes from}

\item{inslides}{Logical. Should the notes be embedded in the slides?
Defaults to \code{TRUE}.}
}
\description{
Function to pull notes from a slide
}
\keyword{internal}
